+++ R/rcpphnsw-package.R
@@ -8,7 +8,8 @@
 #' @docType package
 #' @name RcppHnsw-package
 #' @aliases HnswL2 Rcpp_HnswL2-class HnswCosine Rcpp_HnswCosine-class HnswIp
-#' @aliases Rcpp_HnswIp-class RcppHNSW-package
+#' @aliases Rcpp_HnswIp-class HnswEuclidean Rcpp_HnswEuclidean-class
+#' @aliases RcppHNSW-package
 #' @references
 #' <https://github.com/nmslib/hnswlib>
 #' @author James Melville for the R interface; Yury Malkov for hnswlib itself.
@@ -25,6 +26,7 @@
 Rcpp::loadModule("HnswL2", TRUE)
 Rcpp::loadModule("HnswCosine", TRUE)
 Rcpp::loadModule("HnswIp", TRUE)
+Rcpp::loadModule("HnswEuclidean", TRUE)
 
 .onUnload <- function(libpath) {
   library.dynam.unload("RcppHNSW", libpath)