+++ src/RcppExports.cpp
@@ -14,11 +14,13 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 RcppExport SEXP _rcpp_module_boot_HnswL2();
 RcppExport SEXP _rcpp_module_boot_HnswCosine();
 RcppExport SEXP _rcpp_module_boot_HnswIp();
+RcppExport SEXP _rcpp_module_boot_HnswEuclidean();
 
 static const R_CallMethodDef CallEntries[] = {
     {"_rcpp_module_boot_HnswL2", (DL_FUNC) &_rcpp_module_boot_HnswL2, 0},
     {"_rcpp_module_boot_HnswCosine", (DL_FUNC) &_rcpp_module_boot_HnswCosine, 0},
     {"_rcpp_module_boot_HnswIp", (DL_FUNC) &_rcpp_module_boot_HnswIp, 0},
+    {"_rcpp_module_boot_HnswEuclidean", (DL_FUNC) &_rcpp_module_boot_HnswEuclidean, 0},
     {NULL, NULL, 0}
 };