+++ tests/testthat/test_search_one.R
@@ -22,7 +22,7 @@ expect_equal(index$getNNs(ui10[1, ], 4), self_nn_index4[1, ])
 expect_equal(index$getNNsList(ui10[1, ], 4, FALSE)$item, self_nn_index4[1, ])
 nbrs_with_distances <- index$getNNsList(ui10[1, ], 4, TRUE)
 expect_equal(nbrs_with_distances$item, self_nn_index4[1, ])
-expect_equal(nbrs_with_distances$distance, self_nn_dist4[1, ]^2, tol = 1e-6)
+expect_equal(nbrs_with_distances$distance, self_nn_dist4[1, ], tol = 1e-6)
 expect_error(index$getNNsList(ui10[1, ], 15, FALSE), "(?i)unable to find")
 
 # Test deletion